+++ Content.Client/DeltaV/Overlays/UltraVisionSystem.cs
@@ -1,14 +1,16 @@
+using Content.Shared.Abilities;
+using Content.Shared.DeltaV.CCVars;
 using Robust.Client.Graphics;
-using Content.Shared.Abilities; # DeltaV
-using Content.Shared.DeltaV.CCVars; # DeltaV
-using Robust.Shared.Configuration; # DeltaV
+using Robust.Shared.Configuration;
+using Robust.Shared.Player;
 
 namespace Content.Client.DeltaV.Overlays;
 
 public sealed partial class UltraVisionSystem : EntitySystem
 {
     [Dependency] private readonly IOverlayManager _overlayMan = default!;
-    [Dependency] private readonly IConfigurationManager _cfg = default!; # DeltaV
+    [Dependency] private readonly IConfigurationManager _cfg = default!;
+    [Dependency] private readonly ISharedPlayerManager _playerMan = default!;
 
     private UltraVisionOverlay _overlay = default!;
 
@@ -18,21 +20,35 @@ public sealed partial class UltraVisionSystem : EntitySystem
 
         SubscribeLocalEvent<UltraVisionComponent, ComponentInit>(OnUltraVisionInit);
         SubscribeLocalEvent<UltraVisionComponent, ComponentShutdown>(OnUltraVisionShutdown);
+        SubscribeLocalEvent<UltraVisionComponent, LocalPlayerAttachedEvent>(OnPlayerAttached);
+        SubscribeLocalEvent<UltraVisionComponent, LocalPlayerDetachedEvent>(OnPlayerDetached);
 
-        Subs.CVar(_cfg, DCCVars.NoVisionFilters, OnNoVisionFiltersChanged); # DeltaV
+        Subs.CVar(_cfg, DCCVars.NoVisionFilters, OnNoVisionFiltersChanged);
 
         _overlay = new();
     }
 
     private void OnUltraVisionInit(EntityUid uid, UltraVisionComponent component, ComponentInit args)
     {
-        if (!_cfg.GetCVar(DCCVars.NoVisionFilters)) # DeltaV
-            _overlayMan.AddOverlay(_overlay); # DeltaV
+        if (uid == _playerMan.LocalEntity && !_cfg.GetCVar(DCCVars.NoVisionFilters))
+            _overlayMan.AddOverlay(_overlay);
     }
 
     private void OnUltraVisionShutdown(EntityUid uid, UltraVisionComponent component, ComponentShutdown args)
     {
-        _overlayMan.RemoveOverlay(_overlay); # DeltaV
+        if (uid == _playerMan.LocalEntity)
+            _overlayMan.RemoveOverlay(_overlay);
+    }
+
+    private void OnPlayerAttached(EntityUid uid, UltraVisionComponent component, LocalPlayerAttachedEvent args)
+    {
+        if (!_cfg.GetCVar(DCCVars.NoVisionFilters))
+            _overlayMan.AddOverlay(_overlay);
+    }
+
+    private void OnPlayerDetached(EntityUid uid, UltraVisionComponent component, LocalPlayerDetachedEvent args)
+    {
+        _overlayMan.RemoveOverlay(_overlay);
     }
 
     private void OnNoVisionFiltersChanged(bool enabled)
