+++ Content.Client/Options/UI/OptionsMenu.xaml
@@ -9,6 +9,6 @@
         <tabs:KeyRebindTab Name="KeyRebindTab" />
         <tabs:AudioTab Name="AudioTab" />
         <tabs:NetworkTab Name="NetworkTab" />
-        <dtabs:DeltaTab Name="DeltaTab" />
+        <dtabs:DeltaTab Name="ExtraTab" />
     </TabContainer>
 </DefaultWindow>
