+++ Content.Client/Options/UI/OptionsMenu.xaml.cs
@@ -20,7 +20,7 @@ namespace Content.Client.Options.UI
             Tabs.SetTabTitle(2, Loc.GetString("ui-options-tab-controls"));
             Tabs.SetTabTitle(3, Loc.GetString("ui-options-tab-audio"));
             Tabs.SetTabTitle(4, Loc.GetString("ui-options-tab-network"));
-            Tabs.SetTabTitle(5, Loc.GetString("ui-options-tab-deltav")); // DeltaV specific settings
+            Tabs.SetTabTitle(5, Loc.GetString("ui-options-tab-extra")); // DeltaV specific settings
 
             UpdateTabs();
         }
