--- Content.Server/_NF/VulpLanguage/VulpLangaugeListenerComponent.cs
@@ -1,7 +1,0 @@
-namespace Content.Server.VulpLangauge
-{
-    [RegisterComponent]
-    public partial class VulpLangaugeListenerComponent : Component
-    {
-    }
-}
