--- Content.Server/_NF/VulpLanguage/VulpLanguageSpeakerComponent.cs
@@ -1,7 +1,0 @@
-namespace Content.Server.VulpLangauge
-{
-    [RegisterComponent]
-    public partial class VulpLanguageSpeakerComponent : Component
-    {
-    }
-}
