--- Content.Server/_NF/VulpLanguage/VulpTranslatorComponent.cs
@@ -1,7 +1,0 @@
-namespace Content.Server.VulpLangauge
-{
-    [RegisterComponent]
-    public partial class VulpTranslatorComponent : Component
-    {
-    }
-}
