--- Content.Shared/_NF/VulpLangauge/VulpGiveTranslatorComponent.cs
@@ -1,7 +1,0 @@
-namespace Content.Shared.VulpLangauge
-{
-    [RegisterComponent]
-    public partial class VulpGiveTranslatorComponent : Component
-    {
-    }
-}
