+++ Resources/Locale/en-US/deltav/escape-menu/options-menu.ftl
@@ -1,4 +1,4 @@
-﻿ui-options-tab-deltav = DeltaV
+﻿ui-options-tab-extra = Extra
 ui-options-general-forknotice = Note: These settings are fork-specific and might not apply on other servers.
 
 ui-options-no-filters = Disable species vision filters
