+++ Resources/Prototypes/Catalog/VendingMachines/Inventories/vendomat.yml
@@ -11,5 +11,4 @@
     CapacitorStockPart: 8
     MicroManipulatorStockPart: 8
     Beaker: 8    # Frontier - beakers are required for some machines
-    VulpTranslator: 10
     AirFreshener: 10
