--- Resources/Prototypes/DeltaV/Entities/Objects/Devices/vulptranslator.yml
@@ -1,14 +1,0 @@
-- type: entity
-  id: VulpTranslator
-  parent: [ BaseItem, PowerCellSlotMediumItem ]
-  name: Canilunzt translator
-  description: "Used by Vulpkanins to translate their speech."
-  components:
-  - type: Sprite
-    sprite: /Textures/Objects/Devices/vulp_translator.rsi
-    layers:
-    - state: icon
-  - type: PowerCellDraw
-    drawRate: 0
-    useRate: 1
-  - type: VulpTranslator
