+++ Resources/Prototypes/Nyanotrasen/Entities/Mobs/NPCs/dogs.yml
@@ -55,7 +55,7 @@
     interactSuccessSound:
       path: /Audio/Animals/small_dog_bark_happy.ogg
     hostileOnFail: true
-#  - type: DogVision
+  - type: DogVision
   - type: NPCRetaliation
   - type: FactionException
   - type: NpcFactionMember
