--- Resources/Prototypes/_NF/Traits/disabilities.yml
@@ -1,5 +1,0 @@
-- type: trait
-  id: Colorblindness
-  name: Colorblindness
-  components:
-    - type: DogVision
