+++ Resources/Textures/DeltaV/Effects/creampie.rsi/meta.json
@@ -1,0 +1,1 @@
+{"version": 1, "size": {"x": 32, "y": 32}, "license": "CC-BY-SA-3.0", "copyright": "Edited by Floofers", "states": [{"name": "creampie_vulpkanin", "directions": 4}]}
