--- Resources/Textures/Objects/Devices/vulp_translator.rsi/meta.json
@@ -1,14 +1,0 @@
-{
-    "version": 1,
-    "license": "CC-BY-SA-3.0",
-    "copyright": "baystation12",
-    "size": {
-        "x": 32,
-        "y": 32
-    },
-    "states": [
-        {
-            "name": "icon"
-        }
-    ]
-}
